pagination component 줄이기
