+++ src/components/Pagination.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { PageNum } from 'components';
+import classnames from 'classnames';
 import styles from 'scss/components/Pagination.module.scss';
 
 type Props = {
@@ -8,154 +8,104 @@ type Props = {
   currentPage: number;
   handlePageNum: (value: number) => void;
 };
+type PageNumProps = {
+  isCurrentPage?: boolean;
+  pageNum: number;
+  currentPage?: number;
+  handlePageNum: (value: number) => void;
+};
 
-const UnderPagination = ({ pageAmount, currentPage, handlePageNum }: Props) => {
-  var rows = [];
-  for (var i = 1; i < pageAmount + 1; i++) {
-    rows.push(
-      <PageNum
-        pageNum={i}
-        currentPage={currentPage}
-        handlePageNum={handlePageNum}
-      />,
-    );
+const PageNum: React.FC<PageNumProps> = ({
+  pageNum,
+  isCurrentPage,
+  currentPage,
+  handlePageNum,
+}) => {
+  if (currentPage) {
+    isCurrentPage = currentPage === pageNum;
   }
   return (
-    <>
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage - 1)}
-      >
-        <div>&#60;</div>
-      </div>
-      {rows}
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage + 1)}
-      >
-        <div>&#62;</div>
-      </div>
-    </>
+    <div
+      className={classnames(styles.pageNum, {
+        [styles.currentPage]: isCurrentPage,
+      })}
+      onClick={() => handlePageNum(pageNum)}
+    >
+      {pageNum}
+    </div>
   );
 };
-const SidePagination = ({ pageAmount, currentPage, handlePageNum }: Props) => {
-  const isDots = currentPage < 6;
-  const middleNum = currentPage < 6 ? 5 : pageAmount - 4;
 
-  return (
-    <>
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage - 1)}
-      >
-        <div>&#60;</div>
-      </div>
-      <div className={styles.numList}>
-        <PageNum
-          pageNum={1}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        {!isDots ? (
-          <div className={styles.pageDots}>···</div>
-        ) : (
-          <PageNum
-            pageNum={2}
-            currentPage={currentPage}
-            handlePageNum={handlePageNum}
-          />
-        )}
-
-        <PageNum
-          pageNum={middleNum - 2}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        <PageNum
-          pageNum={middleNum - 1}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        <PageNum
-          pageNum={middleNum}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        <PageNum
-          pageNum={middleNum + 1}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        <PageNum
-          pageNum={middleNum + 2}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-        {isDots ? (
-          <div className={styles.pageDots}>···</div>
-        ) : (
-          <PageNum
-            pageNum={pageAmount - 1}
-            currentPage={currentPage}
-            handlePageNum={handlePageNum}
-          />
-        )}
-        <PageNum
-          pageNum={pageAmount}
-          currentPage={currentPage}
-          handlePageNum={handlePageNum}
-        />
-      </div>
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage + 1)}
-      >
-        <div>&#62;</div>
-      </div>
-    </>
-  );
-};
-const MiddlePagination = ({
+const UpdatedPagination = ({
   pageAmount,
   currentPage,
   handlePageNum,
 }: Props) => {
+  const leftDots = currentPage - 1 > 4;
+
+  const rightDots = pageAmount - currentPage > 4 && pageAmount > 9;
+  const rows = [];
+  const pageNum =
+    currentPage < 6
+      ? 3
+      : pageAmount - currentPage < 5
+      ? pageAmount - 6
+      : currentPage - 2;
+  if (pageAmount < 10) {
+    for (let i = 2; i < pageAmount; i++) {
+      rows.push(
+        <PageNum
+          pageNum={i}
+          currentPage={currentPage}
+          handlePageNum={handlePageNum}
+        />,
+      );
+    }
+  } else {
+    for (let i = pageNum; i < pageNum + 5; i++) {
+      rows.push(
+        <PageNum
+          pageNum={i}
+          currentPage={currentPage}
+          handlePageNum={handlePageNum}
+        />,
+      );
+    }
+  }
+
   return (
     <>
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage - 1)}
-      >
-        <div>&#60;</div>
-      </div>
-      <div className={styles.numList}>
-        <PageNum pageNum={1} handlePageNum={handlePageNum} />
+      {pageAmount < 10 ? (
+        ''
+      ) : leftDots ? (
         <div className={styles.pageDots}>···</div>
-        <PageNum pageNum={currentPage - 2} handlePageNum={handlePageNum} />
-        <PageNum pageNum={currentPage - 1} handlePageNum={handlePageNum} />
+      ) : (
         <PageNum
-          pageNum={currentPage}
-          isCurrentPage={true}
+          pageNum={2}
+          currentPage={currentPage}
           handlePageNum={handlePageNum}
         />
-        <PageNum pageNum={currentPage + 1} handlePageNum={handlePageNum} />
-        <PageNum pageNum={currentPage + 2} handlePageNum={handlePageNum} />
+      )}
+      {rows}
+      {pageAmount < 10 ? (
+        ''
+      ) : rightDots ? (
         <div className={styles.pageDots}>···</div>
-        <PageNum pageNum={pageAmount} handlePageNum={handlePageNum} />
-      </div>
-      <div
-        className={styles.moveNumList}
-        onClick={() => handlePageNum(currentPage + 1)}
-      >
-        <div>&#62;</div>
-      </div>
+      ) : (
+        <PageNum
+          pageNum={pageAmount - 1}
+          currentPage={currentPage}
+          handlePageNum={handlePageNum}
+        />
+      )}
     </>
   );
 };
 
 const Pagination: React.FC = () => {
   const [currentPage, setCurrentPage] = useState<number>(1);
-  const pageAmount = 23;
+  const pageAmount = 13;
+
   //각종 함수들
   const handlePageNum = (value: number) => {
     if (value < 1 || value > pageAmount) {
@@ -163,19 +113,41 @@ const Pagination: React.FC = () => {
     }
     setCurrentPage(value);
   };
-  const StatusPagination =
-    pageAmount < 10
-      ? UnderPagination
-      : 5 < currentPage && currentPage < pageAmount - 5
-      ? MiddlePagination
-      : SidePagination;
+
   return (
     <div className={styles.paginationLayout}>
-      <StatusPagination
+      <div
+        className={styles.moveNumList}
+        onClick={() => handlePageNum(currentPage - 1)}
+      >
+        <div>&#60;</div>
+      </div>
+      <PageNum
+        pageNum={1}
+        currentPage={currentPage}
+        handlePageNum={handlePageNum}
+      />
+
+      <UpdatedPagination
         currentPage={currentPage}
         pageAmount={pageAmount}
         handlePageNum={handlePageNum}
       />
+      {pageAmount < 2 ? (
+        ''
+      ) : (
+        <PageNum
+          pageNum={pageAmount}
+          currentPage={currentPage}
+          handlePageNum={handlePageNum}
+        />
+      )}
+      <div
+        className={styles.moveNumList}
+        onClick={() => handlePageNum(currentPage + 1)}
+      >
+        <div>&#62;</div>
+      </div>
     </div>
   );
 };
