+++ src/components/index.ts
@@ -15,5 +15,4 @@ export { default as Loading } from './Loading';
 export { default as ToastContainer } from './ToastContainer';
 export { default as FetchingCard } from './FetchingCard';
 export { default as Pagination } from './Pagination';
-export { default as PageNum } from './PageNum';
 export { default as CommonCard } from './CommonCard';
